+++ MiaoMiao/app/src/main/java/com/zhangshirong/miaomiao/DataObject/Struct/NoteBookPage.java
@@ -3,7 +3,7 @@ package com.zhangshirong.miaomiao.DataObject.Struct;
 import java.util.ArrayList;
 
 public class NoteBookPage {
-    public int bookId;
-    public int pageId;
+    public long bookId;
+    public long pageId;
     public ArrayList<NoteBookPageRow> rows = new ArrayList<NoteBookPageRow>();
 }
