+++ MiaoMiao/app/src/main/java/com/zhangshirong/miaomiao/DataObject/Struct/NoteBookPageRow.java
@@ -3,7 +3,7 @@ package com.zhangshirong.miaomiao.DataObject.Struct;
 import java.util.ArrayList;
 
 public class NoteBookPageRow{
-    public int pageId;
-    public int rowId;
+    public long pageId;
+    public long rowId;
     public ArrayList<NoteBookPageRowBlock> blocks = new ArrayList<NoteBookPageRowBlock>();
 }
