+++ MiaoMiao/app/src/main/java/com/zhangshirong/miaomiao/DataObject/Struct/NoteBookPageRowBlock.java
@@ -1,8 +1,8 @@
 package com.zhangshirong.miaomiao.DataObject.Struct;
 
 public class NoteBookPageRowBlock {
-    public int blockId;
-    public int rowId;
+    public long blockId;
+    public long rowId;
     public String content;
     public Boolean fillable;
